Use an overlay for custom haskell configuration
This is superior to using packageOverrides as we can compose our local
Haskell project changes with the global overlay definitions which might
exist to fix particular compiler version issues, for example.

As a concrete example, the rebase library doesn't work with ghc821,
and I don't want to have to cater for that problem in each individual
project which has it as a dependency (through hasql). The default
compiler should still work, but if I want to use ghc821, I can still
build the projects using nix, and have the global overlay take care of
the problem.

Also added xmonad and xmobar to my local haskell environment since I
was previously building with stack.