Use callHackage to override nix dependencies
This makes the build more robust against version changes.

When using rebase_1_1_1 etc, the build breaks every time the
version is updated (it's now rebase_1_2_1). By using callHackage
we can lock the environment to a specific version until we
decide to change it, rather than when the nixpkgs channel
updates it.

Not that callHackage relies on the current value of all-cabal-hashes
which is an archive pulled down from the commercialhaskell git repo.
The version of this in nixpkgs is unlikely to be up to date with
the current hackage so may not contain some package versions.
See nixpkgs/pkgs/data/misc/hackage for the current state.

Overriding all-cabal-hashes to use a more recent archive would be
a workaround. Alternatively callCabal2nix could be used to build
an up-to-date version of a dependency, fetched from github.