Disable "proc macro not expanded" rust warnings
I was trying to do this using the wrong settings but got some helpful
advice from matklad on users.rust-lang.org, explaining that I should
add "unresolved-proc-macro" to the list of disabled diagnostics
instead.

In the process I split the LSP config into a separate lua file which
is imported from init.vim.